Refactor refetch evaluation refetch stats to put each stats in their own compoment. The problem we had is that the catch all `useRefetchStats` hook we had was triggering for example numeric mean in non numeric evaluations. The reason is the way SWR stores work by default the store fetch data when it's mounted. This could be solved by disabling fetching on mount but I think is more correct to put each metric a listener to the `evaluationStatus` websocket event
