+++ apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/evaluations/[evaluationId]/_components/MetricsSummary/BigNumberPanels/ModalValuePanel/index.tsx
@@ -1,7 +1,10 @@
 'use client'
 
+import { useCallback } from 'react'
+
 import { EvaluationModalValue } from '@latitude-data/core/browser'
 import { Text } from '@latitude-data/web-ui'
+import { useEvaluationStatusEvent } from '../../../../_lib/useEvaluationStatusEvent'
 import useEvaluationResultsModalValue from '$/stores/evaluationResultCharts/evaluationResultsModalValue'
 
 import Panel from '../Panel'
@@ -17,7 +20,7 @@ export default function ModalValuePanel({
   evaluationId: number
   modal: EvaluationModalValue
 }) {
-  const { data } = useEvaluationResultsModalValue(
+  const { data, refetch } = useEvaluationResultsModalValue(
     {
       commitUuid,
       documentUuid,
@@ -27,6 +30,8 @@ export default function ModalValuePanel({
       fallbackData: modal,
     },
   )
+  const onStatusChange = useCallback(() => refetch(), [refetch])
+  useEvaluationStatusEvent({ evaluationId, documentUuid, onStatusChange })
   return (
     <Panel
       label='Value more repeated'
