+++ apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/evaluations/[evaluationId]/_components/MetricsSummary/BigNumberPanels/TotalsPanels/index.tsx
@@ -1,9 +1,12 @@
 'use client'
 
+import { useCallback } from 'react'
+
 import { EvaluationAggregationTotals } from '@latitude-data/core/browser'
 import { formatCostInMillicents } from '$/app/_lib/formatUtils'
 import useEvaluationResultsCounters from '$/stores/evaluationResultCharts/evaluationResultsCounters'
 
+import { useEvaluationStatusEvent } from '../../../../_lib/useEvaluationStatusEvent'
 import Panel from '../Panel'
 
 export default function TotalsPanels({
@@ -17,7 +20,7 @@ export default function TotalsPanels({
   evaluationId: number
   aggregation: EvaluationAggregationTotals
 }) {
-  const { data } = useEvaluationResultsCounters(
+  const { data, refetch } = useEvaluationResultsCounters(
     {
       commitUuid,
       documentUuid,
@@ -27,6 +30,8 @@ export default function TotalsPanels({
       fallbackData: aggregation,
     },
   )
+  const onStatusChange = useCallback(() => refetch(), [refetch])
+  useEvaluationStatusEvent({ evaluationId, documentUuid, onStatusChange })
   const cost =
     data?.costInMillicents === undefined
       ? '-'
